feat: Add user, who created project and task

+++ ProjectManagment/Application/ProjectTasks/Commands/DeleteTaskCommand.cs
@@ -2,6 +2,7 @@ using Application.Common;
 using Application.Common.Interfaces.Repositories;
 using Application.ProjectTasks.Exceptions;
 using Domain.Tasks;
+using Domain.Users;
 using MediatR;
 
 namespace Application.ProjectTasks.Commands;
@@ -9,29 +10,43 @@ namespace Application.ProjectTasks.Commands;
 public class DeleteTaskCommand : IRequest<Result<ProjectTask, TaskException>>
 {
     public required Guid TaskId { get; init; }
+    public required Guid UserId { get; init; }
 }
 
-public class DeleteTaskCommandHandler(ITaskRepository repository)
+public class DeleteTaskCommandHandler(ITaskRepository repository, IUserRepository userRepository)
     : IRequestHandler<DeleteTaskCommand, Result<ProjectTask, TaskException>>
 {
     public async Task<Result<ProjectTask, TaskException>> Handle(DeleteTaskCommand request,
         CancellationToken cancellationToken)
     {
         var exisitingTask = await repository.GetById(new ProjectTaskId(request.TaskId), cancellationToken);
-
-        return await exisitingTask.Match(
-            async t => await DeleteEntity(t, cancellationToken),
+        var exsistingUser = await userRepository.GetById(new UserId(request.UserId), cancellationToken);
+        return await exsistingUser.Match(
+            async u =>
+            {
+                return await exisitingTask.Match(
+                    async t => await DeleteEntity(t, u, cancellationToken),
+                    () => Task.FromResult<Result<ProjectTask, TaskException>>(
+                        new TaskNotFoundException(new ProjectTaskId(request.TaskId))));
+            },
             () => Task.FromResult<Result<ProjectTask, TaskException>>(
-                new TaskNotFoundException(new ProjectTaskId(request.TaskId))));
+                new UserNotFoundWhileCreated(new ProjectTaskId(request.TaskId)))
+        );
     }
 
     private async Task<Result<ProjectTask, TaskException>> DeleteEntity(
         ProjectTask task,
+        User user,
         CancellationToken cancellationToken)
     {
         try
         {
-            return await repository.Delete(task, cancellationToken);
+            if (user.Id == task.UserId || user.Role!.Name == "Admin")
+            {
+                return await repository.Delete(task, cancellationToken);
+            }
+
+            return new NotEnoughPermission(task.TaskId);
         }
         catch (Exception e)
         {
